requirements.txt
streamlit
streamlit-authenticator
mysql-connector-python
keras
opencv-python
numpy
pytube
yt-dlp